// src/generated/packs.ts
export const ClippyAnimation = {
  gestureLeft: "GestureLeft",
  gestureDown: "GestureDown",
  gestureRight: "GestureRight",
  gestureUp: "GestureUp",
  idleFingerTap: "IdleFingerTap",
  idleSideToSide: "IdleSideToSide",
  idleEyeBrowRaise: "IdleEyeBrowRaise",
  getArtsy: "GetArtsy",
  lookDownRight: "LookDownRight",
  lookDown: "LookDown",
  lookDownLeft: "LookDownLeft",
  lookRight: "LookRight",
  lookLeft: "LookLeft",
  hearing1: "Hearing_1",
  lookUpRight: "LookUpRight",
  lookUp: "LookUp",
  lookUpLeft: "LookUpLeft",
  getAttention: "GetAttention",
  wave: "Wave",
  save: "Save",
  congratulate: "Congratulate",
  processing: "Processing",
  goodBye: "GoodBye",
  print: "Print",
  checkingSomething: "CheckingSomething",
  emptyTrash: "EmptyTrash",
  idleRopePile: "IdleRopePile",
  idleSnooze: "IdleSnooze",
  idleHeadScratch: "IdleHeadScratch",
  sendMail: "SendMail",
  searching: "Searching",
  thinking: "Thinking",
  writing: "Writing",
  explain: "Explain",
  restPose: "RestPose",
  show: "Show",
  hide: "Hide",
  getTechy: "GetTechy",
  getWizardy: "GetWizardy",
  alert: "Alert",
  idleAtom: "IdleAtom",
  greeting: "Greeting",
  idle11: "Idle1_1",
} as const

export const ClippyAnimationMetadata = {
  "GestureLeft": { duration: 3.0500000000000003, loops: true },
  "GestureDown": { duration: 2.2500000000000004, loops: true },
  "GestureRight": { duration: 3.2500000000000004, loops: true },
  "GestureUp": { duration: 2.8000000000000003, loops: true },
  "IdleFingerTap": { duration: 1.1500000000000001, loops: true },
  "IdleSideToSide": { duration: 5.683333333333331, loops: true },
  "IdleEyeBrowRaise": { duration: 1.5000000000000004, loops: true },
  "GetArtsy": { duration: 4.899999999999999, loops: true },
  "LookDownRight": { duration: 1.8000000000000003, loops: true },
  "LookDown": { duration: 1.8000000000000003, loops: true },
  "LookDownLeft": { duration: 1.8000000000000003, loops: true },
  "LookRight": { duration: 1.8000000000000003, loops: true },
  "LookLeft": { duration: 1.8000000000000003, loops: true },
  "Hearing_1": { duration: 5.399999999999997, loops: true },
  "LookUpRight": { duration: 1.8000000000000003, loops: true },
  "LookUp": { duration: 1.8000000000000003, loops: true },
  "LookUpLeft": { duration: 1.8000000000000003, loops: true },
  "GetAttention": { duration: 2.6500000000000004, loops: true },
  "Wave": { duration: 4.9, loops: true },
  "Save": { duration: 5.529999999999997, loops: true },
  "Congratulate": { duration: 3.6800000000000006, loops: true },
  "Processing": { duration: 3.800000000000002, loops: true },
  "GoodBye": { duration: 4.45, loops: true },
  "Print": { duration: 8.399999999999991, loops: true },
  "CheckingSomething": { duration: 6.639999999999995, loops: true },
  "EmptyTrash": { duration: 4.999999999999998, loops: true },
  "IdleRopePile": { duration: 7.499999999999989, loops: true },
  "IdleSnooze": { duration: 13.599999999999989, loops: true },
  "IdleHeadScratch": { duration: 1.9000000000000006, loops: true },
  "SendMail": { duration: 6.799999999999994, loops: true },
  "Searching": { duration: 8.09999999999999, loops: true },
  "Thinking": { duration: 4.5, loops: true },
  "Writing": { duration: 8.399999999999995, loops: true },
  "Explain": { duration: 1.5000000000000004, loops: true },
  "RestPose": { duration: 0.1, loops: true },
  "Show": { duration: 0.05, loops: true },
  "Hide": { duration: 0.05, loops: true },
  "GetTechy": { duration: 4.5, loops: true },
  "GetWizardy": { duration: 3.6800000000000006, loops: true },
  "Alert": { duration: 2.400000000000001, loops: true },
  "IdleAtom": { duration: 4.5, loops: true },
  "Greeting": { duration: 4.45, loops: true },
  "Idle1_1": { duration: 7.299999999999995, loops: true },
} as const

export type ClippyAnimationName = (typeof ClippyAnimation)[keyof typeof ClippyAnimation]

export const CatAnimation = {
  idleButterFly: "IdleButterFly",
  congratulate: "Congratulate",
  idleBlink: "IdleBlink",
  idleYawn: "IdleYawn",
  idleScratch: "IdleScratch",
  greeting: "Greeting",
  deepIdleE: "DeepIdleE",
  goodBye: "GoodBye",
  idleTwitch: "IdleTwitch",
  lookDownRight: "LookDownRight",
  lookDown: "LookDown",
  lookDownLeft: "LookDownLeft",
  lookRight: "LookRight",
  lookLeft: "LookLeft",
  lookUpRight: "LookUpRight",
  lookUp: "LookUp",
  lookUpLeft: "LookUpLeft",
  getAttention: "GetAttention",
  wave: "Wave",
  save: "Save",
  searching: "Searching",
  writing: "Writing",
  idleTailWagA: "IdleTailWagA",
  idleTailWagB: "IdleTailWagB",
  idleTailWagC: "IdleTailWagC",
  idleTailWagD: "IdleTailWagD",
  idleCleaning: "IdleCleaning",
  idleStretch: "IdleStretch",
  emptyTrash: "EmptyTrash",
  idleLegLick: "IdleLegLick",
  getArtsy: "GetArtsy",
  sendMail: "SendMail",
  print: "Print",
  checkingSomething: "CheckingSomething",
  getTechy: "GetTechy",
  deepIdleA: "DeepIdleA",
  explain: "Explain",
  getWizardy: "GetWizardy",
  hearing1: "Hearing_1",
  alert: "Alert",
  gestureDown: "GestureDown",
  gestureLeft: "GestureLeft",
  gestureRight: "GestureRight",
  gestureUp: "GestureUp",
  restPose: "RestPose",
  show: "Show",
  hide: "Hide",
  idle11: "Idle1_1",
  thinking: "Thinking",
  processing: "Processing",
} as const

export const CatAnimationMetadata = {
  "IdleButterFly": { duration: 3.400000000000002, loops: true },
  "Congratulate": { duration: 2.500000000000001, loops: true },
  "IdleBlink": { duration: 0.30000000000000004, loops: true },
  "IdleYawn": { duration: 1.8000000000000005, loops: true },
  "IdleScratch": { duration: 1.5000000000000002, loops: true },
  "Greeting": { duration: 5.799999999999995, loops: true },
  "DeepIdleE": { duration: 14.099999999999977, loops: true },
  "GoodBye": { duration: 1.7000000000000004, loops: true },
  "IdleTwitch": { duration: 1.0999999999999999, loops: true },
  "LookDownRight": { duration: 1.2000000000000002, loops: true },
  "LookDown": { duration: 1.2000000000000002, loops: true },
  "LookDownLeft": { duration: 1.2000000000000002, loops: true },
  "LookRight": { duration: 1.2000000000000002, loops: true },
  "LookLeft": { duration: 1.2000000000000002, loops: true },
  "LookUpRight": { duration: 1.2000000000000002, loops: true },
  "LookUp": { duration: 1.2000000000000002, loops: true },
  "LookUpLeft": { duration: 1.2000000000000002, loops: true },
  "GetAttention": { duration: 4.100000000000001, loops: true },
  "Wave": { duration: 5.949999999999999, loops: true },
  "Save": { duration: 3.900000000000002, loops: true },
  "Searching": { duration: 3.2000000000000006, loops: true },
  "Writing": { duration: 7.939999999999994, loops: true },
  "IdleTailWagA": { duration: 1.4000000000000001, loops: true },
  "IdleTailWagB": { duration: 1.7000000000000004, loops: true },
  "IdleTailWagC": { duration: 1.9500000000000006, loops: true },
  "IdleTailWagD": { duration: 1.6000000000000003, loops: true },
  "IdleCleaning": { duration: 3.5000000000000013, loops: true },
  "IdleStretch": { duration: 3.9000000000000017, loops: true },
  "EmptyTrash": { duration: 3.0000000000000013, loops: true },
  "IdleLegLick": { duration: 2.7999999999999994, loops: true },
  "GetArtsy": { duration: 5.399999999999999, loops: true },
  "SendMail": { duration: 3.5000000000000018, loops: true },
  "Print": { duration: 4.480000000000001, loops: true },
  "CheckingSomething": { duration: 5.599999999999998, loops: true },
  "GetTechy": { duration: 0.9999999999999999, loops: true },
  "DeepIdleA": { duration: 7.599999999999993, loops: true },
  "Explain": { duration: 1.0999999999999999, loops: true },
  "GetWizardy": { duration: 1.4000000000000001, loops: true },
  "Hearing_1": { duration: 1.4000000000000001, loops: true },
  "Alert": { duration: 0.7, loops: true },
  "GestureDown": { duration: 1.2000000000000002, loops: true },
  "GestureLeft": { duration: 1.2000000000000002, loops: true },
  "GestureRight": { duration: 1.2000000000000002, loops: true },
  "GestureUp": { duration: 1.2000000000000002, loops: true },
  "RestPose": { duration: 0.1, loops: true },
  "Show": { duration: 0.04, loops: true },
  "Hide": { duration: 0.04, loops: true },
  "Idle1_1": { duration: 14.903333333333304, loops: true },
  "Thinking": { duration: 0.5, loops: true },
  "Processing": { duration: 5.599999999999998, loops: true },
} as const

export type CatAnimationName = (typeof CatAnimation)[keyof typeof CatAnimation]

export const RockyAnimation = {
  sendMail: "SendMail",
  emptyTrash: "EmptyTrash",
  greeting: "Greeting",
  goodbye: "Goodbye",
  explain: "Explain",
  congratulate: "Congratulate",
  print: "Print",
  gestureRight: "GestureRight",
  lookDown: "LookDown",
  lookDownRight: "LookDownRight",
  lookLeft: "LookLeft",
  lookRight: "LookRight",
  lookUpLeft: "LookUpLeft",
  lookUp: "LookUp",
  lookUpRight: "LookUpRight",
  lookDownLeft: "LookDownLeft",
  getArtsy: "GetArtsy",
  hearing1: "Hearing_1",
  getAttention: "GetAttention",
  wave: "Wave",
  thinking: "Thinking",
  processing: "Processing",
  save: "Save",
  writing: "Writing",
  gestureLeft: "GestureLeft",
  gestureDown: "GestureDown",
  gestureUp: "GestureUp",
  deepIdle1: "DeepIdle1",
  checkingSomething: "CheckingSomething",
  restPose: "RestPose",
  idle8: "Idle(8)",
  idle1: "Idle(1)",
  idle11: "Idle1_1",
  idle3: "Idle(3)",
  idle4: "Idle(4)",
  idle5: "Idle(5)",
  idle6: "Idle(6)",
  idle7: "Idle(7)",
  show: "Show",
  hide: "Hide",
  idle9: "Idle(9)",
  searching: "Searching",
  idle2: "Idle(2)",
  getTechy: "GetTechy",
  alert: "Alert",
  getWizardy: "GetWizardy",
} as const
// ...
export const builtInCharacterPacks = {
  clippy: {
    character: "clippy" as CharacterName,
    displayName: "Clippy",
    atlasUrl: new URL('../assets/Clippy/atlas.png', import.meta.url).href,
    manifestUrl: new URL('../assets/Clippy/manifest.json', import.meta.url).href,
    animationNames: Object.values(ClippyAnimation),
    animationMetadata: ClippyAnimationMetadata,
  },
  cat: {
    character: "cat" as CharacterName,
    displayName: "Cat",
    atlasUrl: new URL('../assets/Cat/atlas.png', import.meta.url).href,
    manifestUrl: new URL('../assets/Cat/manifest.json', import.meta.url).href,
    animationNames: Object.values(CatAnimation),
    animationMetadata: CatAnimationMetadata,
  },
  rocky: {
    character: "rocky" as CharacterName,
    displayName: "Rocky",
    atlasUrl: new URL('../assets/Rocky/atlas.png', import.meta.url).href,
    manifestUrl: new URL('../assets/Rocky/manifest.json', import.meta.url).href,
    animationNames: Object.values(RockyAnimation),
    animationMetadata: RockyAnimationMetadata,
  },
} as const